TableQuestionAnsweringPipeline
